requirements.txt
groq
openai
pandas
numpy
scipy
nltk
textstat
tqdm
matplotlib
seaborn
python-dotenv
jupyter

